Refactored: DAO, TaskService
