+++ src/main/java/ru/javarush/todo/dao/TaskDao.java
@@ -1,18 +1,56 @@
 package ru.javarush.todo.dao;
 
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.query.Query;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 import ru.javarush.todo.entity.Task;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface TaskDao {
-    List<Task> findAll(int offset, int limit);
+@Repository
+public class TaskDao {
+    private final SessionFactory sessionFactory;
 
-    int findAllCount();
+    public TaskDao(SessionFactory sessionFactory) {
+        this.sessionFactory = sessionFactory;
+    }
 
-    Optional<Task> findById(int id);
+    @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
+    public List<Task> findAll(int offset, int limit) {
+        Query<Task> query = getSession().createQuery("select t from Task t", Task.class);
+        query.setFirstResult(offset);
+        query.setMaxResults(limit);
+        return query.getResultList();
+    }
 
-    void saveOrUpdate(Task task);
+    @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
+    public int findAllCount() {
+        Query<Long> query = getSession().createQuery("select count(t) from Task t", Long.class);
+        return Math.toIntExact(query.uniqueResult());
+    }
 
-    void delete(Task task);
+    @Transactional(propagation = Propagation.REQUIRED)
+    public Optional<Task> findById(int id) {
+        Query<Task> query = getSession().createQuery("select t from Task t where t.id = :ID", Task.class);
+        query.setParameter("ID", id);
+        return query.uniqueResultOptional();
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public void saveOrUpdate(Task task) {
+        getSession().persist(task);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public void delete(Task task) {
+        getSession().remove(task);
+    }
+
+    private Session getSession() {
+        return sessionFactory.getCurrentSession();
+    }
 }
